+++ flutter_mobx/test/flutter_mobx_test.dart
@@ -59,6 +59,47 @@ void main() {
     expect(renderCount, equals(1));
   });
 
+  testWidgets("Observer.withChild's child doesn't re-render", (tester) async {
+    final message = Observable('Click');
+     final key1 = UniqueKey();
+     final key2 = UniqueKey();
+     final key3 = UniqueKey();
+
+     await tester.pumpWidget(
+       MaterialApp(
+         home: Observer.withChild(
+           builderWithChild: (context, child) {          
+             return Column(
+               children: [
+                 ElevatedButton(onPressed: () => message.value = 'Clicked', child: Container()),
+                 Text(message.value, key: key1),
+                 child!,
+                 Builder(
+                   builder: (context) {
+                     return Text(message.value, key: key3); 
+                   }
+                 ),
+               ],
+             );
+           },
+           child: Text(message.value, key: key2),
+         ),
+       ),
+     );    
+
+     expect(tester.widget<Text>(find.byKey(key1)).data, equals('Click'));
+     expect(tester.widget<Text>(find.byKey(key2)).data, equals('Click'));
+     expect(tester.widget<Text>(find.byKey(key3)).data, equals('Click'));
+
+     await tester.tap(find.byType(ElevatedButton));
+     expect(message.value, equals('Clicked'));
+
+     await tester.pump();
+     expect(tester.widget<Text>(find.byKey(key1)).data, equals('Clicked')); // Observer rebuilt the Text1
+     expect(tester.widget<Text>(find.byKey(key2)).data, equals('Click')); // child Text2 did not change
+     expect(tester.widget<Text>(find.byKey(key3)).data, equals('Clicked')); // Builder does not preserve from rebuild
+  });
+
   testWidgets('Observer build should call reaction.track', (tester) async {
     final mock = MockReaction();
     when(() => mock.hasObservables).thenReturn(true);